refactor(todos): extract completion-validation orchestration into a bounded module
Addresses the P2 follow-up from #3142's review: move the caller-approved
completion-validation orchestration (read declared command -> run -> gate)
out of the 2318-line loopx/todos.py and into a new bounded module,
loopx/control_plane/todos/completion_validation.py.

complete_goal_todo now calls a single run_completion_validation_gate(...)
before acquiring the mutation lock; the gate, the pre-read, the runner, and
the typed-failure receipt construction all live in the new module. No
behavior change — this is a pure move (the privacy invariant, the
outside-the-lock placement, the dry_run/terminal-replay skip, and the 20s
timeout are all preserved).

Effect: loopx/todos.py 2318 -> 2167 lines; canary ratchet baseline for
loopx/todos.py ratcheted back 2318 -> 2167 accordingly. The new module
follows its control_plane/todos/* siblings (not individually budgeted).

Stacked on #3142 (branch based on feat/todo-completion-validation); rebase
onto main and open as a standalone follow-up PR once #3142 merges.
