+++ loopx/canary/module_metric_baseline.json
@@ -103,7 +103,7 @@
     "loopx/todos.py": {
       "any_count": 48,
       "dict_any_count": 0,
-      "lines": 2318
+      "lines": 2165
     },
     "loopx/worker_bridge.py": {
       "any_count": 28,
